Use configured lancer-initiative module specifier.
This was hardcoded to use "lancer-initiative" to grab flag data, but the
flags are under the system name for the in-system version of
lancer-initiative. By using 'CONFIG.LancerInitiative.module', the
correct flags will be referenced. Future changes should ensure the
lancer-initiative api is used, rather than using raw access to flags.